Fix PlotHeatmap view.
